webnfc/manual-test/nfc_push.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2017 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->
<meta charset=utf-8>
<title>Web NFC Push Test.</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="author" title="Zhao Ming" href="mailto:mingx.zhao@intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/#the-nfc-interface"/>
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<script src="./nfc_help.js"></script>

<div id="log"></div>

<script>
"use strict";
const invalid_messages =
  [
    // NFCMessage.data: should have at least 1 valid record.
    // https://w3c.github.io/web-nfc/#the-push-method - Step 8.
    createMessage([{}]),
    // https://w3c.github.io/web-nfc/#dfn-map-text-to-ndef
    // NFCRecord must have data.
    createMessage([createTextRecord()]),
    // NFCRecord.mediaType for "text" record must be "text/*".
    createMessage([createRecord("text", "application/json",
        test_number_data)]),
    // NFCRecord.data for "text" record must be number or string.
    createMessage([createTextRecord(test_buffer_data)]),
    createMessage([createTextRecord(test_json_data)]),
    // https://w3c.github.io/web-nfc/#dfn-map-a-json-object-to-ndef
    // NFCRecord must have data.
    createMessage([createJsonRecord()]),
    // NFCRecord.data for "json" record must be object.
    createMessage([createJsonRecord(test_buffer_data)]),
    createMessage([createJsonRecord(test_number_data)]),
    createMessage([createJsonRecord(test_text_data)]),
    // NFCRecord.mediaType for "json" record must be "application/json".
    createMessage([createRecord("json", "image/png", test_json_data)]),
    // https://w3c.github.io/web-nfc/#dfn-map-a-url-to-ndef
    // NFCRecord must have data.
    createMessage([createUrlRecord()]),
    // NFCRecord.data for "url" record must be string.
    createMessage([createUrlRecord(test_buffer_data)]),
    createMessage([createUrlRecord(test_number_data)]),
    createMessage([createUrlRecord(test_json_data)]),
    // https://w3c.github.io/web-nfc/#dfn-map-binary-data-to-ndef
    // NFCRecord must have data.
    createMessage([createOpaqueRecord()]),
    // NFCRecord.data for "opaque" record must be ArrayBuffer.
    createMessage([createOpaqueRecord(test_text_data)]),
    createMessage([createOpaqueRecord(test_number_data)]),
    createMessage([createOpaqueRecord(test_json_data)])
  ];

promise_test(t => {
  return promise_rejects(t, "SyntaxError", navigator.nfc.push(undefined));
}, "Test that passing undefined to nfc.push would raise SyntaxError.");

promise_test(t => {
  let promises = [];
  invalid_messages.forEach(message => {
    promises.push(
      return promise_rejects(t, "SyntaxError", navigator.nfc.push(message)));
  });
  Promise.all(promises);
}, "Test that promise is rejected with SyntaxError if NFCMessage is invalid.");
// blocked!

//           // promise_test(t => {
//           //   return navigator.nfc.push(test_text_data, { timeout: 1000 });
//           // }, "nfc.push should succeed when NFC HW is enabled.");
//           // need interactive
//           
//           // promise_test(t => {
//           //   return promise_rejects(t, "TimeoutError", navigator.nfc.push(test_text_data, { timeout: 0}));
//           // }, "nfc.push should fail with TimeoutError when push operation is not completed before specified timeout value.");
  
 promise_test(t => {
   return promise_rejects(t, "SyntaxError", navigator.nfc.push(test_text_data, { timeout: "invalid"}));
 }, "nfc.push should fail with SyntaxError when timeout value is invalid.");

// promise_test(t => {
//   return navigator.nfc.push(test_text_data).then(() => {
//     assertNFCPushOptionsEqual(createNFCPushOptions("any", Infinity, true), test_text_data);
//   });
// }, "Check that default NFCPushOptions values are correctly set.");
// 
// promise_test(t => {
//   let nfcPushOptions = createNFCPushOptions("tag", 1, false);
//   return navigator.nfc.push(test_text_data, nfcPushOptions).then((pushOptions) => {
//     assert_equals(nfcPushOptions, pushOptions);
//   });
// }, "Check that provided NFCPushOptions values are correctly converted.");
// 
// promise_test(t => {
//   return navigator.nfc.push("bbbb")
// }, "bbb.");
// promise_test(t => {
//   return navigator.nfc.push("aaa")
//     .then(() => {
//       navigator.nfc.cancelPush();
//     });
// }, "nfc.cancelPush should succeed if there is pending push operation.");

//promise_test(t => {
//  return navigator.nfc.push("cccc", { timeout :100 });
//  return promise_rejects(t, "AbortError", navigator.nfc.cancelPush());
//}, "nfc.cancelPush should reject pending promise with AbortError.");

promise_test(t => {
  return promise_rejects(t, "NotSupportedError", navigator.nfc.push(new ArrayBuffer(32 * 1024 + 1)));
}, "Reject promise with NotSupportedError if NFC message size exceeds 32KB.");

</script>
